made three-components own prop, fixed some styling
