+++ src/components/TileGrid/TileEdit.tsx
@@ -10,9 +10,12 @@ const TileEditBreadcrumbs = () => (
     strokeMiterlimit="10"
     clipRule="evenodd"
     viewBox="0 0 1000 1000"
-    style={{ transform: "rotate(90deg)", height: "20px" }}
+    style={{
+      transform: "rotate(90deg)",
+      height: "20px",
+    }}
   >
-    <g fill="#5e4b3c">
+    <g fill="#22555A">
       <path d="M115.104 498.827c0-41.24 33.431-74.672 74.671-74.672 41.24 0 74.672 33.432 74.672 74.672 0 41.24-33.432 74.671-74.672 74.671-41.24 0-74.671-33.431-74.671-74.671zM426.622 498.827c0-41.24 33.431-74.672 74.671-74.672 41.24 0 74.672 33.432 74.672 74.672 0 41.24-33.432 74.671-74.672 74.671-41.24 0-74.671-33.431-74.671-74.671zM738.14 498.827c0-41.24 33.432-74.672 74.672-74.672 41.24 0 74.671 33.432 74.671 74.672 0 41.24-33.431 74.671-74.671 74.671-41.24 0-74.672-33.431-74.672-74.671z"></path>
     </g>
   </svg>
@@ -59,7 +62,7 @@ function TileEditMenu({ visible, setVisible, options }: ITileEditMenu) {
     <div
       className={cn(
         visible ? "flex" : "hidden",
-        " absolute z-10 flex-col rounded bg-white text-center drop-shadow-md"
+        "absolute z-10 flex-col rounded bg-white text-center drop-shadow-md"
       )}
       style={{ borderRadius: "6px", paddingInline: "10px" }}
     >
